bug fixed for rating bar size

+++ app/src/main/java/com/example/giftme/Fragments/DetailViewFragment.java
@@ -126,7 +126,7 @@ public class DetailViewFragment extends Fragment {
                         = new Item(Integer.parseInt(cursor.getString(0)),
                         cursor.getString(1),
                         cursor.getString(2),
-                        Integer.parseInt(cursor.getString(3)),
+                        Float.parseFloat(cursor.getString(3)),
                         Integer.parseInt(cursor.getString(4)),
                         cursor.getString(5),
                         cursor.getString(6),
